Live plot: Move uavcan node to file in network
